--- front/src/dashboard/slips/SlipsContainer.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import SlipsHeader from "./SlipsHeader";
-import SlipsContent from "./SlipsContent";
-
-export default function SlipsContainer() {
-  return (
-    <>
-      <SlipsHeader />
-      <SlipsContent />
-    </>
-  );
-}
