--- front/src/dashboard/slips/SlipsHeader.module.scss
@@ -1,3 +1,0 @@
-.slipsHeader {
-  padding: 1rem;
-}
