--- front/src/dashboard/slips/SlipsHeaderActions.module.scss
@@ -1,8 +1,0 @@
-.slipHeaderActions{
-  display: flex;
-  flex-direction: row;
-  padding: 0 1rem 0.75rem 1rem;
-}
-.refreshButton{
-  margin-left: auto;
-}
